remove docs, fix mod level docs, move tests to get_msg_events
