+++ Cargo.toml
@@ -20,6 +20,5 @@ ruma-events = "0.18.0"
 ruma-identifiers = "0.14.1"
 serde = { version = "1.0.105", features = ["derive"] }
 serde_json = "1.0.50"
-serde_urlencoded = "0.6.1"
 strum = { version = "0.18.0", features = ["derive"] }
 url = { version = "2.1.1", features = ["serde"] }
