+++ src/r0/message/get_message_events.rs
@@ -76,3 +76,80 @@ pub enum Direction {
     #[serde(rename = "f")]
     Forward,
 }
+
+#[cfg(test)]
+mod tests {
+    use super::{Direction, Request};
+
+    use std::convert::{TryFrom, TryInto};
+
+    use ruma_identifiers::RoomId;
+
+    use crate::r0::filter::{LazyLoadOptions, RoomEventFilter};
+
+    #[test]
+    fn test_serialize_some_room_event_filter() {
+        let room_id = RoomId::try_from("!roomid:example.org").unwrap();
+        let filter = RoomEventFilter {
+            lazy_load_options: LazyLoadOptions::Enabled {
+                include_redundant_members: true,
+            },
+            rooms: Some(vec![room_id.clone()]),
+            not_rooms: vec!["room".into(), "room2".into(), "room3".into()],
+            not_types: vec!["type".into()],
+            ..Default::default()
+        };
+        let req = Request {
+            room_id,
+            from: "token".into(),
+            to: Some("token2".into()),
+            dir: Direction::Backward,
+            limit: Some(js_int::UInt::min_value()),
+            filter: Some(filter),
+        };
+
+        let request: http::Request<Vec<u8>> = req.try_into().unwrap();
+        assert_eq!(
+            "from=token&to=token2&dir=b&limit=0&filter=%7B%22not_types%22%3A%5B%22type%22%5D%2C%22not_rooms%22%3A%5B%22room%22%2C%22room2%22%2C%22room3%22%5D%2C%22rooms%22%3A%5B%22%21roomid%3Aexample.org%22%5D%2C%22lazy_load_members%22%3Atrue%2C%22include_redundant_members%22%3Atrue%7D",
+            request.uri().query().unwrap()
+        );
+    }
+
+    #[test]
+    fn test_serialize_none_room_event_filter() {
+        let room_id = RoomId::try_from("!roomid:example.org").unwrap();
+        let req = Request {
+            room_id,
+            from: "token".into(),
+            to: Some("token2".into()),
+            dir: Direction::Backward,
+            limit: Some(js_int::UInt::min_value()),
+            filter: None,
+        };
+
+        let request: http::Request<Vec<u8>> = req.try_into().unwrap();
+        assert_eq!(
+            "from=token&to=token2&dir=b&limit=0",
+            request.uri().query().unwrap(),
+        );
+    }
+
+    #[test]
+    fn test_serialize_default_room_event_filter() {
+        let room_id = RoomId::try_from("!roomid:example.org").unwrap();
+        let req = Request {
+            room_id,
+            from: "token".into(),
+            to: Some("token2".into()),
+            dir: Direction::Backward,
+            limit: Some(js_int::UInt::min_value()),
+            filter: Some(RoomEventFilter::default()),
+        };
+
+        let request: http::Request<Vec<u8>> = req.try_into().unwrap();
+        assert_eq!(
+            "from=token&to=token2&dir=b&limit=0&filter=%7B%7D",
+            request.uri().query().unwrap(),
+        );
+    }
+}
