+++ src/serde/json_string.rs
@@ -1,12 +1,10 @@
-//! De-/serialization functions for `Option<RoomEventFilter>` objects.
-//! Delegates to `js_int::UInt` to ensure integer size is within bounds.
+//! De-/serialization functions to and from json strings, allows the type to be used as a query string.
 
 use serde::{
     de::{Deserialize, DeserializeOwned, Deserializer, Error as _},
     ser::{Error as _, Serialize, Serializer},
 };
 
-/// Serialize a filter into a json string.
 pub fn serialize<T, S>(filter: T, serializer: S) -> Result<S::Ok, S::Error>
 where
     T: Serialize,
@@ -16,7 +14,6 @@ where
     serializer.serialize_str(&json)
 }
 
-/// Deserializes a filter from a json string.
 pub fn deserialize<'de, T, D>(deserializer: D) -> Result<T, D::Error>
 where
     T: DeserializeOwned,
@@ -24,71 +21,4 @@ where
 {
     let s = String::deserialize(deserializer)?;
     serde_json::from_str(&s).map_err(D::Error::custom)
-}
-
-#[cfg(test)]
-mod tests {
-    use std::convert::{TryFrom, TryInto};
-
-    use ruma_identifiers::RoomId;
-
-    use crate::r0::filter::{LazyLoadOptions, RoomEventFilter};
-    use crate::r0::message::get_message_events::{Direction, Request};
-
-    #[test]
-    fn test_serialize_some_room_event_filter() {
-        let room_id = RoomId::try_from("!roomid:example.org").unwrap();
-        let filter = RoomEventFilter {
-            lazy_load_options: LazyLoadOptions::Enabled {
-                include_redundant_members: true,
-            },
-            rooms: Some(vec![room_id.clone()]),
-            not_rooms: vec!["room".into(), "room2".into(), "room3".into()],
-            not_types: vec!["type".into()],
-            ..Default::default()
-        };
-        let req = Request {
-            room_id,
-            from: "token".into(),
-            to: Some("token2".into()),
-            dir: Direction::Backward,
-            limit: Some(js_int::UInt::min_value()),
-            filter: Some(filter),
-        };
-
-        let request: http::Request<Vec<u8>> = req.try_into().unwrap();
-        println!("{:?}", request.uri().query());
-    }
-
-    #[test]
-    fn test_serialize_none_room_event_filter() {
-        let room_id = RoomId::try_from("!roomid:example.org").unwrap();
-        let req = Request {
-            room_id,
-            from: "token".into(),
-            to: Some("token2".into()),
-            dir: Direction::Backward,
-            limit: Some(js_int::UInt::min_value()),
-            filter: None,
-        };
-
-        let request: http::Request<Vec<u8>> = req.try_into().unwrap();
-        println!("{:?}", request.uri());
-    }
-
-    #[test]
-    fn test_serialize_default_room_event_filter() {
-        let room_id = RoomId::try_from("!roomid:example.org").unwrap();
-        let req = Request {
-            room_id,
-            from: "token".into(),
-            to: Some("token2".into()),
-            dir: Direction::Backward,
-            limit: Some(js_int::UInt::min_value()),
-            filter: Some(RoomEventFilter::default()),
-        };
-
-        let request: http::Request<Vec<u8>> = req.try_into().unwrap();
-        println!("{:?}", request.uri());
-    }
 }
